test(server): add tests for photo schema resolver
